fix, feat, refactor, chore according to comments on PR
fix: content no longer overflows poll area on screen resize.
feat: pressing enter in input field allows password submission.
refactor: showResults section moved to separate component. NextBossPoll dictates the visibility of PollArea and PollResults.
chore: moved contants, types out of function declaration.
