+++ client/src/components/NextBossPoll.module.css
@@ -1,16 +1,16 @@
-.pollArea {
+.pollContainer {
   display: flex;
   flex-direction: column;
   width: 100%;
-  height: 27rem;
+  overflow: hidden;
+  max-height: 30rem;
   background-color: var(--ctp-mantle);
   border-radius: 1.5rem;
   gap: 0.2rem;
-  padding-bottom: 1rem;
 }
 
 .titleBox {
-  margin: 1rem 1rem 0.25rem 1rem;
+  margin: 1rem 1rem 0 1rem;
   text-align: center;
   overflow-wrap: break-word;
 }
@@ -24,251 +24,6 @@
   text-transform: uppercase;
 }
 
-.legend {
-  display: flex;
-  direction: row;
-  gap: 5%;
-  justify-content: center;
-  margin: 0 1rem 0 1rem;
-}
-
-.legendToday {
-  color: var(--ctp-peach);
-  text-align: center;
-}
-
-.legendAvailable {
-  color: var(--ctp-sky);
-  text-align: center;
-}
-
-.todayHeader,
-.availableHeader,
-.availableTodayHeader,
-.unavailableHeader {
-  display: flex;
-  font-size: clamp(0.5rem, 0.9vw, 1rem);
-  width: 100%;
-  height: auto;
-  aspect-ratio: 2;
-}
-
-@media (width < 600px) {
-  .todayHeader,
-  .availableHeader,
-  .availableTodayHeader,
-  .unavailableHeader {
-    font-size: 1rem;
-  }
-}
-
-.todayHeader {
-  background-color: var(--ctp-peach);
-  color: var(--ctp-surface0);
-}
-
-.availableHeader {
-  background-color: var(--ctp-sky);
-  color: var(--ctp-surface0);
-  position: relative;
-}
-
-.availableTodayHeader {
-  width: 100%;
-  clip-path: polygon(0 100%, 100% 0, 100% 100%);
-  transform: translateY(0.5px);
-  will-change: transform;
-  background-color: var(--ctp-peach);
-  color: var(--ctp-surface0);
-  position: absolute;
-}
-
-.unavailableHeader {
-  background-color: var(--ctp-surface0);
-}
-
-.dayTimeSelectionGrid {
-  display: grid;
-  grid-template-columns: repeat(7, 1fr);
-  box-sizing: border-box;
-  background-color: var(--ctp-mantle);
-  gap: 3px;
-  margin: 0 1rem 0 1rem;
-  overflow-y: auto;
-  cursor: pointer;
-  -moz-user-select: none;
-  user-select: none;
-}
-
-.dayGridColumn {
-  display: grid;
-  grid-template-rows: repeat(11, 1fr);
-}
-
-.timeGridItemSelectable,
-.timeGridItemDisabled {
-  display: grid;
-  grid-template-columns: subgrid;
-  border-bottom: 1px dashed var(--ctp-surface0);
-  font-size: 0.7rem;
-  padding-left: 0.1rem;
-}
-
-.timeGridItemSelectable {
-  background-color: var(--ctp-overlay0);
-}
-
-.timeGridItemSelectable[selected] {
-  background-color: var(--ctp-green);
-  color: var(--ctp-surface0);
-}
-
-.timeGridItemDisabled {
-  background-color: var(--ctp-surface1);
-}
-
-.buttonBar {
-  display: grid;
-  grid-template-columns: 5fr 6fr;
-  padding: 1rem 1.5rem 0 1.5rem;
-  gap: 1rem;
-}
-
-.userSelectContainer {
-  grid-column: 1;
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-start;
-  text-align: center;
-}
-
-.userSelectDropdown {
-  position: relative;
-}
-
-.userSelectDropdownButton {
-  width: 100%;
-  background-color: var(--ctp-surface1);
-  text-align: center;
-  cursor: pointer;
-}
-
-.userSelectDropdownButton .submitButton {
-  height: 2.25rem;
-  padding: 0.25vw 1.5vw 0.25vw 1.5vw;
-}
-
-.userSelectDropdownButton:hover,
-.userSelectDropdownButton:focus {
-  background-color: var(--ctp-surface0);
-  border: 1px solid var(--ctp-overlay1);
-}
-
-.userSelectDropdownButton[clicked] {
-  border-bottom-left-radius: 0;
-  border-bottom-right-radius: 0;
-}
-
-.userSelectDropdownButton[clicked] + .userSelectDropdownContent {
-  display: block;
-}
-
-.userSelectDropdownContent {
-  display: none;
-  width: 100%;
-  max-height: 7.5rem;
-  overflow-y: auto;
-  scrollbar-width: none;
-  -ms-overflow-style: none;
-  border-bottom-left-radius: 8px;
-  border-bottom-right-radius: 8px;
-  position: absolute;
-  cursor: pointer;
-}
-
-.userSelectDropdownContent::-webkit-scrollbar {
-  width: 0;
-  height: 0;
-}
-
-.userSelectDropdownContent span {
-  display: block;
-  width: inherit;
-  text-align: center;
-  background-color: var(--ctp-surface1);
-  border-top: 1px solid var(--ctp-overlay0);
-  padding: 0.1rem 0 0.1rem 0;
-}
-
-.userSelectDropdownContent span:hover {
-  background-color: var(--ctp-surface0);
-}
-
-input {
-  height: 1rem;
-  width: calc(100% - 4px);
-  background-color: var(--ctp-surface1);
-  border: 2px inset var(--ctp-overlay1);
-  border-radius: 8px;
-  padding: 0.25rem 0 0.25rem 0;
-  text-align: center;
-}
-
-input:hover,
-input:focus {
-  background-color: var(--ctp-surface0);
-}
-
-input:focus,
-input:focus-visible {
-  outline: 1px solid var(--ctp-overlay1);
-}
-
-.submissionArea {
-  grid-column: 2;
-  display: flex;
-  flex-direction: column;
-}
-
-.toolTipArea {
-  display: flex;
-  padding-bottom: 0.5rem;
-  align-items: center;
-}
-
-.questionMark {
-  flex: none;
-  width: 1.2rem;
-  height: 1.2rem;
-  font-family: monospace;
-  font-size: 1rem;
-  text-align: center;
-  align-content: flex-start;
-  line-height: normal;
-  border: 1px solid var(--ctp-surface1);
-  border-radius: 50%;
-  cursor: pointer;
-}
-
-.toolTip {
-  display: inline-block;
-  visibility: hidden;
-  max-height: 5rem;
-  padding-left: 0.5rem;
-  font-size: 0.7rem;
-  -moz-user-select: none;
-  user-select: none;
-}
-
-.questionMark:hover {
-  border: 1px solid var(--ctp-red);
-  filter: blur(0.1rem);
-}
-
-.questionMark:hover ~ .toolTip {
-  visibility: visible;
-}
-
 .submitButton {
   text-align: center;
   color: var(--ctp-crust);
@@ -279,87 +34,4 @@ input:focus-visible {
 
 .submitButton:hover {
   background-color: var(--ctp-sky);
-}
-
-.badFieldTip {
-  display: block;
-  visibility: hidden;
-  text-align: center;
-  color: var(--ctp-red);
-  font-size: clamp(0.7rem, 0.7vw, 1rem);
-}
-
-.badFieldTip[show] {
-  visibility: visible;
-}
-
-.rankingsContainer {
-  display: flex;
-  flex-direction: column;
-  gap: 1.2rem;
-  align-items: center;
-  padding-top: 0.5rem;
-  overflow: auto;
-}
-
-.rankingsGrid {
-  display: grid;
-  width: 80%;
-  grid-template-rows: repeat(6, 1fr);
-  box-sizing: border-box;
-  background-color: var(--ctp-mantle);
-  gap: 3px;
-}
-
-.rankingsRow {
-  display: grid;
-  grid-template-columns: 4fr 9fr 4fr;
-  gap: 1px;
-}
-
-.rankingsLeftHeader,
-.rankingsRightHeader {
-  background-color: var(--ctp-sky);
-  color: var(--ctp-surface0);
-  margin-bottom: 1px;
-}
-
-.rankingsLoneHeader {
-  background-color: var(--ctp-peach);
-  color: var(--ctp-surface0);
-  margin-bottom: 1px;
-}
-
-.rankingsEntryLone,
-.rankingsEntryLeft,
-.rankingsEntryRight {
-  background-color: var(--ctp-surface0);
-}
-
-.rankingsEntryLone,
-.rankingsLoneHeader {
-  text-align: center;
-  margin-left: 0.1rem;
-  border-radius: 5px;
-}
-
-.rankingsEntryLeft,
-.rankingsLeftHeader {
-  text-align: center;
-  border-top-left-radius: 5px;
-  border-bottom-left-radius: 5px;
-}
-
-.rankingsEntryRight,
-.rankingsRightHeader {
-  text-align: left;
-  padding-left: 1rem;
-  border-top-right-radius: 5px;
-  border-bottom-right-radius: 5px;
-}
-
-.buttonContainer {
-  display: flex;
-  align-items: center;
-  margin-right: 1rem;
 }
