+++ client/src/components/PollResults.module.css
@@ -1,0 +1,73 @@
+.rankingsContainer {
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+  align-items: center;
+  padding-top: 0.5rem;
+  overflow: auto;
+}
+
+.rankingsGrid {
+  display: grid;
+  width: 80%;
+  grid-template-rows: repeat(6, 1fr);
+  box-sizing: border-box;
+  background-color: var(--ctp-mantle);
+  gap: 3px;
+}
+
+.rankingsRow {
+  display: grid;
+  grid-template-columns: 4fr 9fr 4fr;
+  gap: 1px;
+}
+
+.rankingsLeftHeader,
+.rankingsRightHeader {
+  font-size: calc(0.6rem + 0.4vw);
+  background-color: var(--ctp-sky);
+  color: var(--ctp-surface0);
+  margin-bottom: 1px;
+}
+
+.rankingsLoneHeader {
+  font-size: calc(0.6rem + 0.4vw);
+  background-color: var(--ctp-peach);
+  color: var(--ctp-surface0);
+  margin-bottom: 1px;
+}
+
+.rankingsEntryLone,
+.rankingsEntryLeft,
+.rankingsEntryRight {
+  font-size: calc(0.6rem + 0.4vw);
+  background-color: var(--ctp-surface0);
+}
+
+.rankingsEntryLone,
+.rankingsLoneHeader {
+  text-align: center;
+  margin-left: 0.1rem;
+  border-radius: 5px;
+}
+
+.rankingsEntryLeft,
+.rankingsLeftHeader {
+  text-align: center;
+  border-top-left-radius: 5px;
+  border-bottom-left-radius: 5px;
+}
+
+.rankingsEntryRight,
+.rankingsRightHeader {
+  text-align: left;
+  padding-left: 1rem;
+  border-top-right-radius: 5px;
+  border-bottom-right-radius: 5px;
+}
+
+.buttonContainer {
+  display: flex;
+  align-items: center;
+  margin: 0 0 1.5rem 0;
+}
